moves timer out to standalone compoent
